start migrating to TypeScript
