+++ .github/workflows/test.yml
@@ -24,5 +24,8 @@ jobs:
       - name: Install dependencies
         run: npm i
 
+      - name: Build
+        run: npm run build
+
       - name: Run tests
         run: npm test
