+++ index.test.ts
@@ -1,14 +1,14 @@
 /* eslint-disable no-useless-escape */
 // Idk how useless the escapes are here, need to check later.
-const test = require('ava').default;
+import test from 'ava';
 
-const fs = require('fs');
+import fs from 'fs';
 
-const { paths } = require('./src/config');
-const { envParse, envRead, readFileOrReturnEmptyObject } = require('./src/envParseRead');
-const { envStringifyInit } = require('./src/envStringifyInit');
-const { setAll } = require('./src/setAll');
-const { upsertFile } = require('./src/upsertFile');
+import { paths } from './src/config';
+import { envParse, envRead, readFileOrReturnEmptyObject } from './src/envParseRead';
+import { envStringifyInit } from './src/envStringifyInit';
+import { setAll } from './src/setAll';
+import { upsertFile } from './src/upsertFile';
 
 const envTestData = {
   schema: `MAIL_USERNAME=^\S+$
@@ -17,16 +17,16 @@ MAIL_PASSWORD=^\S+$
 
 MAIL_PORT=^\S+$
 GOOGLE_API_KEY=^\S+$
-BOT_TOKEN=^\S+$ 
+BOT_TOKEN=^\S+$
  # testing for comments = lol
 BOT_OWNER_ID=^[0-9]+$
 MONGO=(ftp|http(s)|mongodb):\/\/(\w+:{0,1}\w*@)?(\S+)(:[0-9]+)?(\/|\/([\w#!:.?+=&%@!\-/]))?
 
 # Testing support for comments`,
-  defaults: `MAIL_USERNAME=  
+  defaults: `MAIL_USERNAME=
 
 MAIL_HOST=imap.gmail.com
-MAIL_PORT=993  
+MAIL_PORT=993
 
 
 MONGO=mongodb://localhost:27017/$npm_package_name`,
@@ -58,6 +58,7 @@ test('reads/writes files', (t) => {
     fs.rmSync(paths.env);
   }
   const envSchema = envRead(paths.schema);
+  // @ts-expect-error todo
   setAll(envSchema, '');
 
   const envDefaults = envRead(paths.defaults);
