--- jsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "checkJs": true,
-    "lib": ["ES2021"],
-    "resolveJsonModule": true,
-    "noUnusedLocals": true
-  }
-}
