+++ src/config.ts
@@ -1,9 +1,5 @@
-const paths = {
+export const paths = {
   env: '.env',
   schema: '.env.schema',
   defaults: '.env.defaults',
-};
-
-module.exports = {
-  paths,
 };
