--- src/filterObject.js
@@ -1,9 +1,0 @@
-const filterObject = (obj, predicate) => !!obj && Object.keys(obj).forEach((k) => {
-  if (!predicate(obj[k])) {
-    delete obj[k];
-  }
-});
-
-module.exports = {
-  filterObject,
-};
