--- src/setAll.js
@@ -1,5 +1,0 @@
-const setAll = (obj, val) => !!obj && Object.keys(obj).forEach((k) => { obj[k] = val; });
-
-module.exports = {
-  setAll,
-};
